Added asyncProfileImage as replacement for loadableProfileImage.
